Commit code truoc khi merge

+++ resources/views/admin/ageLimit/index.blade.php
@@ -10,7 +10,8 @@
                 <div class="card-header d-flex justify-content-between align-items-center gap-2">
                     <h4 class="card-title flex-grow-1">Danh sách giới hạn độ tuổi</h4>
 
-                    <form id="delete-selected-age-limit-form" method="POST" action="{{ route('admin.age_limits.bulkDelete') }}">
+                   @can('delete age limit')
+                        <form id="delete-selected-age-limit-form" method="POST" action="{{ route('admin.age_limits.bulkDelete') }}">
                         @csrf
                         <input type="hidden" name="ids" id="selected-age-limit-ids">
                         <button type="submit" class="btn btn-danger btn-sm">
@@ -18,10 +19,13 @@
                             Xóa đã chọn
                         </button>
                     </form>
+                   @endcan
 
-                    <a href="{{ route('admin.age_limits.create') }}" class="btn btn-sm btn-primary">
+                    @can('create age limit')
+                        <a href="{{ route('admin.age_limits.create') }}" class="btn btn-sm btn-primary">
                         Thêm mới
-                    </a>
+                        </a>
+                    @endcan
                 </div>
 
                 <div class="table-responsive">
@@ -55,6 +59,7 @@
                                             <a href="{{ route('admin.age_limits.edit', $ageLimit->id) }}" class="btn btn-sm btn-soft-primary">
                                                 <iconify-icon icon="solar:pen-2-broken" class="align-middle fs-18"></iconify-icon>
                                             </a>
+                                            @can('delete age limit')
                                             <form action="{{ route('admin.age_limits.destroy', $ageLimit->id) }}" method="POST" onsubmit="return confirm('Bạn có chắc muốn xóa?')">
                                                 @csrf
                                                 @method('DELETE')
@@ -62,6 +67,7 @@
                                                     <iconify-icon icon="solar:trash-bin-trash-broken" class="align-middle fs-18"></iconify-icon>
                                                 </button>
                                             </form>
+                                            @endcan
                                         </div>
                                     </td>
                                 </tr>
